fix: resolve clippy undocumented_unsafe_blocks and lint errors
- Add SAFETY comments to unsafe impls/blocks in xgrammar-sys, hip-kernels,
  autograd examples, cuda-kernels tensor
- Add crate-level #![allow(clippy::undocumented_unsafe_blocks)] to cli and
  cuda-kernels/ffi/gemm_tests (test-only FFI call sites)
- Fix needless_range_loop allow in kv_quant test, collapsible_if in paged_kv,
  redundant_closure in tensor, unnecessary_cast in fused_linear_distill
- Fix dead_code and needless_range_loop in infer-cuda qwen35 probe tests

+++ crates/autograd/src/ops/fused_linear_distill.rs
@@ -1683,10 +1683,7 @@ mod tests {
         let hidden_data = deterministic_vec(seq * hidden_dim, 7);
         let weight_data = deterministic_vec(vocab * hidden_dim, 11);
         let positions: Vec<i32> = vec![0, 2, 3, 5, 6];
-        let targets: Vec<i32> = positions
-            .iter()
-            .map(|&p| ((p * 5 + 3) % 32) as i32)
-            .collect();
+        let targets: Vec<i32> = positions.iter().map(|&p| (p * 5 + 3) % 32).collect();
         let n = positions.len();
 
         let rollout = model_target_logprobs(
